% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empathy_long.R
\name{Info1h_l}
\alias{Info1h_l}
\title{Calculate the expected utility in information set 1}
\usage{
Info1h_l(pars)
}
\arguments{
\item{pars}{vector with length of 5}

\item{X}{The benefit from acting}

\item{k}{The cost of inquiry}

\item{h}{The negative utility from showing empathy}

\item{p}{The probability of bad state}

\item{alpha}{The probability of correct inquiry}
}
\value{
The optimal utility in Info Set 1
}
\description{
Calculate the expected utility in information set 1
}
\examples{
X <- 5;
k <- 1;
h <- 10;
p <- 0.4;
alpha <- 0.8;
parameter <- c(X, k, h, p, alpha);
Info1h_l(parameter);
UInfo1h_l <- as.numeric(capture.output(Info1h_l(parameter))[1]);
cat("Utility under optimal strategy (Info Set 1):", UInfo1h_l, "\n")
DInfo1h_l <- as.numeric(capture.output(Info1h_l(parameter))[2]);
cat("U(act) - U(no act) (Info Set 1):", DInfo1h_l, "\n")

#'
X_list <- seq(0.00, 10, by=.01);
alpha_list <- seq(0.5, 1, by = .001);
X <- X_list;
k <- k_list;
h <- 10;
p <- 0.5;
alpha <- 0.8;
matrixW = matrix(data=NA, nrow=length(X_list), ncol=length(k_list));
for(i in 1:length(X_list)){
  for(j in 1:length(k_list)){
    X <- X_list[i]
    k <- k_list[j]
    parameter <- c(X, k, h, p, alpha)
    DInfo1h_l <- as.numeric(capture.output(Info1h_l(parameter))[2])
    matrixW[i,j] <- DInfo1h_l
  }
}
}
