% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empathy_short.R
\name{Info1h_s}
\alias{Info1h_s}
\title{Calculate the expected utility in information set 1}
\usage{
Info1h_s(pars)
}
\arguments{
\item{X}{The benefit from acting}

\item{k}{The cost of inquiry}

\item{h}{The negative utility from showing empathy}

\item{p}{The probability of bad state}

\item{alpha}{The probability of correct inquiry}
}
\value{
The optimal utility in Info Set 1
}
\description{
Calculate the expected utility in information set 1
}
\examples{
u1 <- Info1h_l(c(1, 2, 3, 0.4, 0.7));
}
