% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punishment_long.R
\name{Info1T_l}
\alias{Info1T_l}
\title{Calculate the expected utility in information set 1}
\usage{
Info1T_l(pars)
}
\arguments{
\item{X}{The benefit from acting}

\item{k}{The cost of inquiry}

\item{T0}{The punishment when acting knowingly}

\item{p}{The probability of bad state}

\item{alpha}{The probability of correct inquiry}

\item{T1}{The punishment of willful ignorance}
}
\value{
The optimal utility in Info Set 1
}
\description{
Calculate the expected utility in information set 1
}
\examples{
u1 <- Info1T_l(c(1, 2, 3, 0.4, 0.7, 3));
}
