% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empathy_short.R
\name{Info2h_s}
\alias{Info2h_s}
\title{Calculate utility w/o action in information set 2, that is, if the Dm does not inquire}
\usage{
Info2h_s(pars)
}
\arguments{
\item{X}{The benefit from acting}

\item{k}{The cost of inquiry}

\item{h}{The negative utility from showing empathy}

\item{p}{The probability of bad state}

\item{alpha}{The probability of correct inquiry}
}
\value{
The optimal utility in Info Set 2
}
\description{
Calculate utility w/o action in information set 2, that is, if the Dm does not inquire
}
\examples{
u1 <- Info2h_l(c(1, 2, 3, 0.4, 0.7));
}
