% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punishment_long.R
\name{Info2T_l}
\alias{Info2T_l}
\title{Calculate utility w/o action in information set 2, that is the DO does not inqurie}
\usage{
Info2T_l(pars)
}
\arguments{
\item{pars}{vector with length of 6}

\item{X}{The benefit from acting}

\item{k}{The cost of inquiry}

\item{T0}{The punishment when acting knowingly}

\item{p}{The probability of bad state}

\item{alpha}{The probability of correct inquiry}

\item{T1}{The punishment of willful ignorance}
}
\value{
The optimal utility in Info Set 2
}
\description{
Calculate utility w/o action in information set 2, that is the DO does not inqurie
}
\examples{
X <- 5;
k <- 1;
T0 <- 10;
p <- 0.4;
alpha <- 0.8;
T1 <- 5;
parameter <- c(X, k, T0, p, alpha, T1);
Info2T_l(parameter);
UInfo2T_l <- as.numeric(capture.output(Info2T_l(parameter))[1]);
cat("Utility under optimal strategy (Info Set 2):", UInfo2T_l, "\n")
DInfo2T_l <- as.numeric(capture.output(Info2T_l(parameter))[2]);
cat("U(act) - U(no act) (Info Set 2):", DInfo2T_l, "\n")
}
