% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punishment_long.R
\name{Info3T_l}
\alias{Info3T_l}
\title{Calculate utility w/o action in information set 3, that is, after receiving a bad signal}
\usage{
Info3T_l(pars)
}
\arguments{
\item{X}{The benefit from acting}

\item{k}{The cost of inquiry}

\item{T0}{The punishment when acting knowingly}

\item{p}{The probability of bad state}

\item{alpha}{The probability of correct inquiry}

\item{T1}{The punishment of willful ignorance}
}
\value{
The optimal utility in Info Set 3
}
\description{
Calculate utility w/o action in information set 3, that is, after receiving a bad signal
}
\examples{
u1 <- Info3T_l(c(1, 2, 3, 0.4, 0.7, 3));
}
