% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punishment_short.R
\name{Info3T_s}
\alias{Info3T_s}
\title{Calculate utility w/o action in information set 3, that is, after receiving a bad signal}
\usage{
Info3T_s(pars)
}
\arguments{
\item{pars}{vector with length of 6}

\item{X}{The benefit from acting}

\item{k}{The cost of inquiry}

\item{T0}{The punishment when acting knowingly}

\item{p}{The probability of bad state}

\item{alpha}{The probability of correct inquiry}

\item{T1}{The punishment of willful ignorance}
}
\value{
The optimal utility in Info Set 3
}
\description{
Calculate utility w/o action in information set 3, that is, after receiving a bad signal
}
\examples{
X <- 5;
k <- 1;
T0 <- 10;
p <- 0.4;
alpha <- 0.8;
T1 <- 5;
parameter <- c(X, k, T0, p, alpha, T1);
Info3T_s(parameter);
UInfo3T_s <- as.numeric(capture.output(Info3T_s(parameter))[1]);
cat("Utility under optimal strategy (Info Set 3):", UInfo3T_s, "\n")
DInfo3T_s <- as.numeric(capture.output(Info3T_s(parameter))[2]);
cat("U(act) - U(no act) (Info Set 3):", DInfo3T_s, "\n")
}
