% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empathy_long.R
\name{Info4h_l}
\alias{Info4h_l}
\title{Calculate utility w/o action in information set 4, that is, after receiving a good signal}
\usage{
Info4h_l(pars)
}
\arguments{
\item{X}{The benefit from acting}

\item{k}{The cost of inquiry}

\item{h}{The negative utility from showing empathy}

\item{p}{The probability of bad state}

\item{alpha}{The probability of correct inquiry}
}
\value{
The optimal utility in Info Set 4
}
\description{
Calculate utility w/o action in information set 4, that is, after receiving a good signal
}
\examples{
u1 <- Info4h_l(c(1, 2, 3, 0.4, 0.7));
}
