% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empathy_short.R
\name{Info4h_s}
\alias{Info4h_s}
\title{Calculate utility w/o action in information set 4, that is, after receiving a good signal}
\usage{
Info4h_s(pars)
}
\arguments{
\item{pars}{vector with length of 5}

\item{X}{The benefit from acting}

\item{k}{The cost of inquiry}

\item{h}{The negative utility from showing empathy}

\item{p}{The probability of bad state}

\item{alpha}{The probability of correct inquiry}
}
\value{
The optimal utility in Info Set 4
}
\description{
Calculate utility w/o action in information set 4, that is, after receiving a good signal
}
\examples{
X <- 5;
k <- 1;
h <- 10;
p <- 0.4;
alpha <- 0.8;
parameter <- c(X, k, h, p, alpha);
Info4h_s(parameter);
UInfo4h_s <- as.numeric(capture.output(Info4h_s(parameter))[1]);
cat("Utility under optimal strategy (Info Set 4):", UInfo4h_s, "\n")
DInfo4h_s <- as.numeric(capture.output(Info4h_s(parameter))[2]);
cat("U(act) - U(no act) (Info Set 4):", DInfo4h_s, "\n")
}
