% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punishment_short.R
\name{Info4T_s}
\alias{Info4T_s}
\title{Calculate utility w/o action in information set 4, that is, after receiving a good signal}
\usage{
Info4T_s(pars)
}
\arguments{
\item{X}{The benefit from acting}

\item{k}{The cost of inquiry}

\item{T0}{The punishment when acting knowingly}

\item{p}{The probability of bad state}

\item{alpha}{The probability of correct inquiry}

\item{T1}{The punishment of willful ignorance}
}
\value{
The optimal utility in Info Set 4
}
\description{
Calculate utility w/o action in information set 4, that is, after receiving a good signal
}
\examples{
u1 <- Info3T_l(c(1, 2, 3, 0.4, 0.7, 3));
}
